backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
django-ninja==1.3.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
